formatted time and split out location into offset and primary

+++ app/src/main/java/com/example/android/quakereport/Earthquake.java
@@ -1,9 +1,12 @@
 package com.example.android.quakereport;
 
+import java.text.SimpleDateFormat;
+import java.util.Date;
+
 public class Earthquake {
     private String mLocation;
     private Double mMag;
-    private int mDate;
+    private long mDate;
 
     /**
      * Create location object with no img
@@ -11,7 +14,7 @@ public class Earthquake {
      * @param mag magnitude of the earthquake
      * @param date date of the earthquake
      */
-    public Earthquake(String location, Double mag, int date) {
+    public Earthquake(String location, Double mag, long date) {
         this.mLocation = location;
         this.mMag = mag;
         this.mDate = date;
@@ -34,10 +37,10 @@ public class Earthquake {
     }
 
     /**
-     * date of the earthquake
-     * @return String
+     * date of the earthquake in milliseconds
+     * @return long
      */
-    public int getDate() {
+    public long getDate() {
         return mDate;
     }
 }
